ofi_net: Cleanup of ofi_bsock_flush*() functions
The functions ofi_bsock_flush() and ofi_bsock_flush_async() return
ssize_t, but possible return values are <= 0 only. We can convert
the return type to int and let the function return 0 on success and
< 0 on failure.

The patch also removes a useless "else {}" block.

+++ include/ofi_net.h
@@ -452,8 +452,8 @@ static inline size_t ofi_bsock_tosend(struct ofi_bsock *bsock)
 	return ofi_byteq_readable(&bsock->sq);
 }
 
-ssize_t ofi_bsock_flush(struct ofi_bsock *bsock);
-ssize_t ofi_bsock_flush_sync(struct ofi_bsock *bsock);
+int ofi_bsock_flush(struct ofi_bsock *bsock);
+int ofi_bsock_flush_sync(struct ofi_bsock *bsock);
 /* For sends started asynchronously, the return value will be -EINPROGRESS_ASYNC,
  * and len will be set to the number of bytes that were queued.
  */
